feat(hub-common): aggregations and dyanmic facets
