+++ packages/common/test/search/_internal/state.test.ts
@@ -13,6 +13,8 @@ import {
   applySingleSelectFacetState,
   serializeMultiSelectFacetState,
   serializeSingleSelectFacetState,
+  serializeSortState,
+  serializeFacetState,
 } from "../../../src/search/_internal";
 
 const singleSelectFacet: IFacet = {
@@ -236,10 +238,10 @@ describe("state utils", () => {
     });
   });
   describe("serializeSingleSelectFacetState", () => {
-    it("returns empty object if no selected", () => {
+    it("returns key:null if no selected", () => {
       const f = cloneObject(singleSelectFacet);
       const chk = serializeSingleSelectFacetState(f);
-      expect(chk).toEqual({});
+      expect(chk).toEqual({ tags: null });
     });
     it("returns key with selected entry key", () => {
       const f = cloneObject(singleSelectFacet);
@@ -251,10 +253,10 @@ describe("state utils", () => {
     });
   });
   describe("serializeMultiSelectFacetState", () => {
-    it("returns empty object if no selected", () => {
+    it("returns key:null if no selected", () => {
       const f = cloneObject(multiSelectFacet);
       const chk = serializeMultiSelectFacetState(f);
-      expect(chk).toEqual({});
+      expect(chk).toEqual({ tags: null });
     });
     it("returns key with selected entry key", () => {
       const f = cloneObject(multiSelectFacet);
@@ -263,6 +265,66 @@ describe("state utils", () => {
       const chk = serializeMultiSelectFacetState(f);
       expect(chk).toEqual({
         tags: "river,lake",
+      });
+    });
+  });
+
+  describe("serializeSortState", () => {
+    it("returns string from a sortOption", () => {
+      const opt: ISortOption = {
+        label: "Created",
+        attribute: "created",
+        defaultOrder: "asc",
+        order: "asc",
+      };
+      const chk = serializeSortState(opt);
+      expect(chk).toEqual("Created|created|asc");
+    });
+  });
+
+  describe("serializeFacetState:", () => {
+    describe("single-select:", () => {
+      it("returns empty object if no selected", () => {
+        const f = cloneObject(singleSelectFacet);
+        const chk = serializeFacetState(f);
+        expect(chk).toEqual({ tags: null });
+      });
+      it("returns key with selected entry key", () => {
+        const f = cloneObject(singleSelectFacet);
+        f.options[1].selected = true;
+        const chk = serializeFacetState(f);
+        expect(chk).toEqual({
+          tags: "lake",
+        });
+      });
+    });
+    describe("multi-select:", () => {
+      it("returns empty object if no selected", () => {
+        const f = cloneObject(multiSelectFacet);
+        const chk = serializeFacetState(f);
+        expect(chk).toEqual({ tags: null });
+      });
+      it("returns key with selected entry key", () => {
+        const f = cloneObject(multiSelectFacet);
+        f.options[0].selected = true;
+        f.options[1].selected = true;
+        const chk = serializeFacetState(f);
+        expect(chk).toEqual({
+          tags: "river,lake",
+        });
+      });
+    });
+    describe("other types", () => {
+      it("returns empty state for other types", () => {
+        const f1 = cloneObject(multiSelectFacet);
+        f1.display = "date-range";
+        expect(serializeFacetState(f1)).toEqual({});
+        f1.display = "histogram";
+        expect(serializeFacetState(f1)).toEqual({});
+        const other = {
+          display: "other",
+        } as unknown as IFacet;
+        expect(serializeFacetState(other)).toEqual({});
       });
     });
   });
